+++ CHANGELOG.md
@@ -12,6 +12,15 @@ When we make [non-breaking changes](https://developer.paddle.com/api-reference/a
 
 This means when upgrading minor versions of the SDK, you may notice type errors. You can safely ignore these or fix by adding additional type guards.
 
+## 1.10.0 - 2024-11-13
+
+### Added
+
+- Added `paymentMethods` resources
+- Added `generateAuthToken` for customer
+
+---
+
 ## 1.9.1 - 2024-10-16
 
 ### Fixed